requirements.txt
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.2
colorama==0.4.3
distlib==0.3.0
Django==3.0.7
django-cleanup==5.0.0
django-crispy-forms==1.9.1
django-etc==1.3.0
django-hitcount==1.3.3
django-resized==0.3.11
django-tinymce==3.2.0
filelock==3.0.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
virtualenv==20.0.23
wrapt==1.12.1
